src/ui/requirements.txt
rich
python-dotenv
tk
